Update cone code for newer three.js
Updated the cone code for r87 of three.js. Unfortunately, this required expanding the attribute arrays to all be per-vertex which results in a moderate expansion:
	positions, colors, radius increase from 2 values to 6 per cone
	normals, uv coords have no change

These values can be reduced at a later date by passing them into the shader using a flag to mark it as child or parent.