Modified the build / test process.
Now the build process only uses babel to transpile the code that is
placed into the dist folder. This keeps it simple for inclusion into
other projects. The webpack code is used for testing and places all the
output into the test folder. Now to load up a test window, you need to
run the following:

npm run test
npm serve