Turns off render for every RequestAnimationFrame call.
There is now an option to disable the render call, unless the camera has
been repositioned. This options is useful if the contents of the scene
are static and don't change by themselves. It drastically cuts down on
the amount of rendering that needs to be done by the browser. However
the option is there to turn it on if the items in the scene will are
expected to have some sort of independent motion.