Converts loadCompartment to loadCompartmentFromURL
The original loadCompartment function required the SharkViewer object to
be set up with a url, that was used to load in the compartment data. The
new loadCompartment function takes the raw obj text as an argument and
uses that to load into the scene. This provides more options to the
developer when it comes to fetching the data. For instance, it made it
possible to add an upload 'obj' button to the test site. The original
function was renamed to loadCompartmentFromURL and now expects the full
url to be present, in order to operate.