Now builds sharkViewer as module or library.
This updates the build process to allow sharkViewer to be used as a
module or a library. The rationale is to allow existing users to
continue to include the code in their sites, without having to learn the
intricacies of building modules with npm, etc. It also allows new
developers the chance to include it in their code bundle, to simplify
deployment.