Adds limits on the camera travel distance.
The default max distance the camera can be zoomed out in the
OrbitControls module is infinity. This was way too far and lead to the
neurons being clipped as the user zoomed out. The new limit is set to
100000 by default, but can be overridden. This prevents the camera from
moving too far out and always leaves the neurons in view.