Forces webGL1 context to prevent rendering issues.
After moving from three r87 to r125, some "bubbles" started appearing
in the imposters. This appears to be caused by a switch from using the
webGL1 context to the webGL2 context by default. @hubbarp noticed that
in the migration notes for r118, a new webGL1Renderer was added to force
the use of webGL1. This seems to fix the issue.

https://github.com/mrdoob/three.js/wiki/Migration-Guide#r117--r118